Fix job race condition by atomically creating reading in database
Now, if a reading for this hour has already been created in the database, the job returns early without error. The previous ReadingExistsInHour was decoupled from the insertion of the reading, so a race condition was possible.

The downside of this approach is that the reading is not created until recorder.RecordReading is called, which is after usage has been read. That's a lot of wasted work.