Represent currency as microcredits and support pricing over time
Microcredits: After more reading, I think choosing a specific sub-credit unit and representing it with bigint will make our lives easier. I picked microcredits, each 1e6 of a credit, because the smallest-divisible thing we currently plan to bill for is SES, which is 1 credit per 150k emails. That's 6.666 repeating credits. (Maybe we'll change the pricing to make it something more even, but it's the right order of magnitude.)

Why bigint? We can be explicit with our rounding strategy (round()? floor()?); it is a fixed-size integer datatype (8 bytes) which allows for fast calculations, whereas numeric is variable-length and slower. Overall I think it will make calculations faster and more predictable.

Price table: Prices for resource kinds can change over time. Those changes should be append-only; the record of previous pricing should remain for auditability. Rows in the price table are valid for specific periods of time.