Fix missing timezone causing update bounds to be incorrect; other bugs
Readings are inserted into the database with created_at set to a UTC time by the Go app. However, the column type is `timestamp`, which drops the timezone. The bounds function relies on timezone to know which readings to update. As explained in comment, we cannot update the field to `timestamptz` without dropping the unique index. Instead, we supplement with a generated column that includes the timezone, but no index.

This commit also tests the BoundsMonthPrev function and fixed other bugs with the update_measurement_credits function found in testing.