Define prices per unit
Before this change, the price table was designed assuming the amount_microcredits was for a single "unit" of resource consumption. For instance, a database instance would be measured with value = 1, and the credits consumed for that measurement would be 1 * amount_microcredits. However, many of our prices are defined on units other than our hourly measurement interval. Most prices are in credits per month, where a month is normalized to 730 hours. To account for this, this commit adds a "unit" column to the price table; the microcredits corresponding due a measurement are determined by (measurement value * price amount / price units).