Simplify getting a JWT and drop option to disable auth
Having an option to disable auth is inherently risky. Instead, since we already have UAA client credentials for the billing service in docker.env so the service can pull usage data from CAPI, we can use those credentials to get a JWT with the needed `usage.admin` scope, and use it in requests.