Check accounts and entries exist after test
This is a step towards providing inputs to setup the test and outputs to check the result. assertDBContains still needs some work to test for the presence of other db types.

Simplify commit/rollback. I previously thought t.Cleanup was sometimes failing to run the rollback or commit, so I returned them so they could be called with defer. The actual problem is that I was not checking the errors returned from cleanup and commit, and they were silently failing. I've simplified the method signature so callers can change one values (commit bool) to commit the results of the test.